refactor: reorganize components following Scope Rule with hierarchical folders
- Group related components in folders when a component depends on another
- status-chart/: Contains status-chart.tsx + status-chart.ssr.tsx wrapper
- risk-severity-chart/: Contains risk-severity-chart.tsx + risk-severity-chart.ssr.tsx wrapper
- finding-severity-over-time/: Contains finding-severity-over-time.tsx + finding-severity-over-time.ssr.tsx wrapper
- Add barrel exports (index.ts) for clean imports
- Simple standalone components (accounts-selector, provider-type-selector, time-range-selector) remain at root
- Update page.tsx imports to use barrel exports
