+++ ui/app/(prowler)/new-overview/components/finding-severity-over-time/finding-severity-over-time.tsx
