+++ ui/app/(prowler)/new-overview/components/finding-severity-over-time/index.ts
@@ -1,0 +1,2 @@
+export { FindingSeverityOverTime } from "./finding-severity-over-time";
+export { FindingSeverityOverTimeSSR } from "./finding-severity-over-time.ssr";
