+++ ui/app/(prowler)/new-overview/components/risk-severity-chart/index.ts
@@ -1,0 +1,2 @@
+export { RiskSeverityChart } from "./risk-severity-chart";
+export { RiskSeverityChartSSR } from "./risk-severity-chart.ssr";
