+++ ui/app/(prowler)/new-overview/components/risk-severity-chart/risk-severity-chart.tsx
