+++ ui/app/(prowler)/new-overview/components/status-chart/index.ts
@@ -1,0 +1,2 @@
+export { StatusChart } from "./status-chart";
+export { StatusChartSSR } from "./status-chart.ssr";
