+++ ui/app/(prowler)/new-overview/components/status-chart/status-chart.tsx
